one-pager/banner/requirements.txt
functions-framework
google-cloud-storage
google-auth
google-genai
Pillow